% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_structures.R
\name{as_adj_list1}
\alias{as_adj_list1}
\title{Adjacency list}
\usage{
as_adj_list1(g)
}
\arguments{
\item{g}{An igraph object}
}
\value{
A list of numeric vectors.
}
\description{
Create adjacency lists from a graph, either for adjacent edges or for neighboring vertices. This version is faster than the version of igraph but less general.
}
\details{
The function does not have a mode parameter and returns the same neighbors as \code{as_adj_list(g, mode = "all")}, as plain integer vectors. For directed graphs, both in- and out-neighbors are returned.
}
\examples{
library(igraph)
g <- make_ring(10)
as_adj_list1(g)
}
\author{
David Schoch
}
