% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringCoef.R
\name{clusteringCoef}
\alias{clusteringCoef}
\title{Clustering coefficient}
\usage{
clusteringCoef(net)
}
\arguments{
\item{net}{network object}
}
\value{
clustering coefficient, \deqn{\frac{3 \times triangles}{two-paths}}
}
\description{
Clustering coefficient
}
\examples{
n <- makeNetwork(10, .3)
clusteringCoef(n)
}

