% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseMultiNet.R
\name{collapseMultiNet}
\alias{collapseMultiNet}
\title{Make a simple network with edge attribute from a multiplex network}
\usage{
collapseMultiNet(net)
}
\arguments{
\item{net}{Network}
}
\value{
Network with edge attribute "NumberTies"
}
\description{
Make a simple network with edge attribute from a multiplex network
}
\examples{
collapseMultiNet(simMultiNet(10, 1e3))
collapseMultiNet(simMultiNet(10, 1e3, directed = FALSE))
}
