% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_periphery.R
\name{core_periphery}
\alias{core_periphery}
\title{Discrete core-periphery model}
\usage{
core_periphery(graph, method = "SA", iter = 50000)
}
\arguments{
\item{graph}{igraph object.}

\item{method}{algorithm to use (see details)}

\item{iter}{number of iterations for SA optimization}
}
\value{
list with numeric vector with entries (k1,k2,...ki...) where ki assigns vertex i to either the core (ki=1) or periphery (ki=0), and the maximal correlation with an optimal pattern matrix
}
\description{
Fits a discrete core-periphery model to a give network
}
\details{
The function fits the data to an optimal pattern matrix with simulated annealing (method="SA") or a rank 1 approximation (method="rk1"). The rank 1 approximation is computationally far cheaper but also more experimental. Best is to compare the results from both models.
}
\references{
Borgatti, Stephen P., and Martin G. Everett. "Models of core/periphery structures." Social networks 21.4 (2000): 375-395.
}
\author{
David Schoch
}
