% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaGWD.R
\name{deltaGWD}
\alias{deltaGWD}
\title{Title delta GWD}
\usage{
deltaGWD(k, theta_s)
}
\arguments{
\item{k}{Node degree}

\item{theta_s}{Decay parameter value}
}
\value{
Change to GWD statistic for adding a half-edge to a node of degree-k
}
\description{
Title delta GWD
}
\examples{
deltaGWD(1:10, 1.5)
}

