% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDegreeDist.R
\name{getDegreeDist}
\alias{getDegreeDist}
\title{Title Get degree distributions}
\usage{
getDegreeDist(graphs)
}
\arguments{
\item{graphs}{list of networks}
}
\value{
data.frame with degree, times observed in each graph, and graph number
}
\description{
Title Get degree distributions
}
\examples{
getDegreeDist(replicate(5, makeNetwork(10, .1)))
}

