% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_kpartite}
\alias{graph_kpartite}
\title{k partite graphs}
\usage{
graph_kpartite(n = 10, grp = c(5, 5))
}
\arguments{
\item{n}{number of nodes}

\item{grp}{vector of partition sizes. Must sum to \code{n}.}
}
\value{
igraph object with vertex attribute \code{type} giving the partition of each vertex
}
\description{
Create a complete k-partite graph.
}
\examples{
# 3-partite graph with equal sized groups
graph_kpartite(n = 15, grp = c(5, 5, 5))
}
\author{
David Schoch
}
