% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_to_sage}
\alias{graph_to_sage}
\title{convert igraph object to sage format}
\usage{
graph_to_sage(g)
}
\arguments{
\item{g}{igraph object}
}
\value{
sage string
}
\description{
convert igraph object to sage format to be read in SAGE
}
\author{
David Schoch
}
