% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helpers}
\alias{helpers}
\alias{biggest_component}
\alias{delete_isolates}
\title{helper function}
\usage{
biggest_component(g)

delete_isolates(g)
}
\arguments{
\item{g}{igraph object.}
}
\value{
igraph object
}
\description{
small functions to deal with typical network problems
}
\author{
David Schoch
}
