% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchyMacdonald.R
\name{hierarchyMacdonald}
\alias{hierarchyMacdonald}
\title{Title Hierarchy (MacDonald)}
\usage{
hierarchyMacdonald(net)
}
\arguments{
\item{net}{network, probably directed}
}
\value{
hierarchy in [0, 1]
}
\description{
Title Hierarchy (MacDonald)
}
\details{
Note that this hierarchy is quite different from Krackhardt's or reciprocal, both of which are available in sna::hierarchy().
}
\examples{
hierarchyMacdonald(makeNetwork(5, .1, dir = TRUE))
}
\references{
MacDonald 1983 *Trees and Networks in Biological Models*
}
