% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMultiNet.R
\name{simMultiNet}
\alias{simMultiNet}
\title{Simulate multiple-edges network}
\usage{
simMultiNet(n, e, directed = TRUE, loops = FALSE)
}
\arguments{
\item{n}{Number of nodes}

\item{e}{Number of edges}

\item{directed}{Boolean, only directed currently supported}

\item{loops}{Boolean, whether to allow self-ties}
}
\value{
network object with multiple = TRUE
}
\description{
Simulate multiple-edges network
}
\examples{
simMultiNet(5, 15)
}

