% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_igraph.R
\name{str.igraph}
\alias{str.igraph}
\title{Print graphs to terminal}
\usage{
\method{str}{igraph}(object, ...)
}
\arguments{
\item{object}{An igraph object}

\item{...}{additional arguments to print (ignored)}
}
\description{
Prints an igraph object to terminal (different than the standard igraph method)
}
\author{
David Schoch
}
