% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strongConnectedness.R
\name{strongConnectedness}
\alias{strongConnectedness}
\title{Title strong connectedness}
\usage{
strongConnectedness(net)
}
\arguments{
\item{net}{network (can be sociomatrix)}
}
\value{
fraction of dyads that are connected on directed paths
}
\description{
Title strong connectedness
}
\details{
sna::connectedness is based on weak reachability; that is, dyads are connected even if it takes going against the direction of links. This requires a link *from* i *to* j to consider i-j connected.
}
\examples{
n <- makeNetwork(10, .1, directed = TRUE)
strongConnectedness(n)
}
