% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vAttrDF.R
\name{vAttrDF}
\alias{vAttrDF}
\title{Vertex attributes as a data.frame}
\usage{
vAttrDF(net)
}
\arguments{
\item{net}{}
}
\value{
data.frame with vertex.names as first column, vertex.attributes as rest
}
\description{
Vertex attributes as a data.frame
}

