% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDegreeDist.R
\name{getDegreeDist}
\alias{getDegreeDist}
\title{Get degree distributions}
\usage{
getDegreeDist(graphs)
}
\arguments{
\item{graphs}{list of networks, as from ergm::simulate}
}
\value{
data.frame with degree, times observed in each graph, and graph number
}
\description{
Get degree distributions
}
\details{
Note that `degDist()` is for a single network. This returns the distribution of degree distributions for a stack of graphs, typically from a call to ergm::simulate with nsim > 1.
}
\examples{
getDegreeDist(replicate(5, makeNetwork(10, .1), simplify = FALSE))
}

