% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_products.R
\name{graph_cartesian}
\alias{graph_cartesian}
\title{Cartesian product of two graphs}
\usage{
graph_cartesian(g, h)
}
\arguments{
\item{g}{An igraph object}

\item{h}{An igraph object}
}
\value{
Cartesian product as igraph object
}
\description{
Compute the Cartesian product of two graphs
}
\details{
See https://en.wikipedia.org/wiki/Cartesian_product_of_graphs
}
\examples{
library(igraph)
g <- make_ring(4)
h <- make_full_graph(2)
graph_cartesian(g, h)
}
\author{
David Schoch
}
