% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qap.R
\name{graph_cor}
\alias{graph_cor}
\alias{graph_cor.default}
\alias{graph_cor.igraph}
\alias{graph_cor.matrix}
\alias{graph_cor.array}
\title{Graph correlation}
\usage{
graph_cor(object1, object2)

\method{graph_cor}{default}(object1, object2)

\method{graph_cor}{igraph}(object1, object2, ...)

\method{graph_cor}{matrix}(object1, object2)

\method{graph_cor}{array}(object1, object2)
}
\arguments{
\item{object1}{igraph object or adjacency matrix}

\item{object2}{igraph object or adjacency matrix over the same vertex set as object1}

\item{...}{additional arguments}
}
\value{
correlation between graphs
}
\description{
This function computes the correlation between networks. Implemented methods expect
the graph to be an adjacency matrix, an igraph, or a network object.
}
