% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_products.R
\name{graph_direct}
\alias{graph_direct}
\title{Direct product of two graphs}
\usage{
graph_direct(g, h)
}
\arguments{
\item{g}{An igraph object}

\item{h}{An igraph object}
}
\value{
Direct product as igraph object
}
\description{
Compute the direct product of two graphs
}
\details{
See https://en.wikipedia.org/wiki/Tensor_product_of_graphs
}
\author{
David Schoch
}
