% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_to_sage}
\alias{graph_to_sage}
\title{convert igraph object to sage format}
\usage{
graph_to_sage(g)
}
\arguments{
\item{g}{igraph object}
}
\value{
sage string
}
\description{
convert igraph object to sage format
}
\details{
sage code

gis = g.is_interval(certificate=true)

gis

gisstr=str(gis)

o = open('interval_raw.txt','w')

o.write(gisstr)

o.close()

system('grep -oe "\link{0-9}\\+,\\s\link{0-9}\\+" interval_raw.txt > intervals.txt')
}
\author{
David Schoch
}
