% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMPL.R
\name{HMPL}
\alias{HMPL}
\title{Harmonic mean path length}
\usage{
HMPL(net)
}
\arguments{
\item{net}{network object. if `net` is directed, will be made undirected by the weak rule.}
}
\value{
Harmonic mean path length between pairs of nodes.
}
\description{
Harmonic mean path length
}
\examples{
n <- network(rgraph(10, tprob = .1))
HMPL(n)
}
\references{
Newman MEJ, 2003. The structure and function of complex networks. SIAM Review.
}

