% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNetwork.R
\name{makeNetwork}
\alias{makeNetwork}
\title{Title Make a network ob with (stochastic) density}
\usage{
makeNetwork(nSize, nDensity, directed = FALSE)
}
\arguments{
\item{nSize}{Number of nodes}

\item{nDensity}{Probability of each edge}
}
\value{
network object
}
\description{
Title Make a network ob with (stochastic) density
}
\examples{
makeNetwork(10, .05)
makeNetwork(10, .05, TRUE)
}

