% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{split_graph}
\alias{split_graph}
\title{split graph}
\usage{
split_graph(n, p, core)
}
\arguments{
\item{n}{number of nodes}

\item{p}{probability of peripheral nodes to connect to the core nodes}

\item{core}{fraction of nodes in the core}
}
\value{
igraph object
}
\description{
Create a random split graph with a perfect core-periphery structure.
}
\author{
David Schoch
}
