% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triad_census_attr.R
\name{triad_census_attr}
\alias{triad_census_attr}
\title{triad census with node attributes}
\usage{
triad_census_attr(g, vattr)
}
\arguments{
\item{g}{igraph object. should be a directed graph.}

\item{vattr}{name of vertex attribute to be used.}
}
\value{
triad census with node attributes.
}
\description{
triad census with node attributes
}
\details{
The node attribute should be integers from 1 to max(attr).
}
\author{
David Schoch
}
